+++ src/types/aiStyle.js
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=aiStyle.js.map